+++ webapi/webapi/Repositories/GamesRepository.cs
@@ -7,7 +7,7 @@ public class GamesRepository(AppDbContext db)
 {
 	private readonly AppDbContext db = db;
 
-	public Task<long> GetIdByName(string gameName)
+	public Task<int> GetIdByName(string gameName)
 	{
 		return db.Games.Where(x => x.Name == gameName).Select(x => x.ID).FirstAsync();
 	}
